LayoutTests/http/tests/cookies/same-site/fetch-after-navigating-iframe-in-cross-origin-page.https-expected.txt



--------
Frame: '<!--frame1-->'
--------
Tests that Same-Site cookies for 127.0.0.1 are not sent with a frame navigation for a frame embedded in a page with a different origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Cookies sent with HTTP request:
PASS Do not have cookie "strict".
PASS Do not have cookie "lax".
PASS Has cookie "none".
PASS defaultSameSiteValueCookieExists is invalidSameSiteValueCookieExists

Cookies visible in DOM:
PASS Do not have DOM cookie "strict".
PASS Do not have DOM cookie "lax".
PASS Has DOM cookie "none".
PASS defaultSameSiteValueDOMCookieExists is invalidSameSiteValueDOMCookieExists
PASS defaultSameSiteValueDOMCookieExists is defaultSameSiteValueCookieExists
PASS successfullyParsed is true

TEST COMPLETE
